Chore: OAuthManager -> GithubLoginManager로 이름 변경
